refactor: split chat room
